Got user land programs to work again!

+++ src/libc/include/stddef.h
@@ -5,7 +5,7 @@
 #endif
 
 typedef long long ptrdiff_t;
-typedef long unsigned int size_t;
+typedef unsigned long long size_t;
 typedef char wchar_t;
 
 #define offsetof(st, m) ((size_t)&(((st*)0)->m))
